+++ src/helpers/index.js
@@ -18,6 +18,13 @@ import {
   updateLastTweetId,
   addMissingTweets,
   removeMissingTweets,
+  addGiftUser,
+  removeGiftUser,
+  isGiftUser,
+  isMessageRecipient,
+  addGiftMessage,
+  removeGiftMessage,
+  getGiftMessage,
 } from "./dbHelper.js";
 
 import {
@@ -69,4 +76,11 @@ export {
   updateLastTweetId,
   addMissingTweets,
   removeMissingTweets,
+  addGiftUser,
+  removeGiftUser,
+  isGiftUser,
+  isMessageRecipient,
+  addGiftMessage,
+  removeGiftMessage,
+  getGiftMessage,
 };